Move shared functions into main_test
