+++ sql/main_test.go
@@ -18,10 +18,13 @@
 package sql_test
 
 import (
+	"database/sql"
 	"testing"
 
+	"github.com/cockroachdb/cockroach/client"
 	"github.com/cockroachdb/cockroach/security"
 	"github.com/cockroachdb/cockroach/security/securitytest"
+	"github.com/cockroachdb/cockroach/server"
 	"github.com/cockroachdb/cockroach/util/leaktest"
 )
 
@@ -33,4 +36,23 @@ func init() {
 
 func TestMain(m *testing.M) {
 	leaktest.TestMainWithLeakCheck(m)
+}
+
+func setup(t *testing.T) (*server.TestServer, *sql.DB, *client.DB) {
+	s := server.StartTestServer(nil)
+	sqlDB, err := sql.Open("cockroach", "https://root@"+s.ServingAddr()+"?certs=test_certs")
+	if err != nil {
+		t.Fatal(err)
+	}
+	kvDB, err := client.Open("https://root@" + s.ServingAddr() + "?certs=test_certs")
+	if err != nil {
+		t.Fatal(err)
+	}
+
+	return s, sqlDB, kvDB
+}
+
+func cleanup(s *server.TestServer, db *sql.DB) {
+	_ = db.Close()
+	s.Stop()
 }
